# chore: upgrade Next.js to 14.2.25 and remove --legacy-peer-deps