--- a/jest.e2e.config.js
+++ b/jest.e2e.config.js
@@ -1,0 +1,26 @@
+module.exports = {
+  preset: 'jest-puppeteer',
+  testMatch: ['<rootDir>/e2e/**/*.test.{js,ts,tsx}'],
+  testTimeout: 120000,
+  setupFilesAfterEnv: ['<rootDir>/e2e/setup.ts'],
+  testEnvironment: 'jest-environment-jsdom',
+  transform: {
+    '^.+\\.(ts|tsx)$': ['babel-jest', { presets: ['next/babel'] }],
+  },
+  moduleNameMapper: {
+    '^@/(.*)$': '<rootDir>/src/$1',
+  },
+  collectCoverageFrom: [
+    'src/components/**/*.{ts,tsx}',
+    '!src/components/**/*.d.ts',
+    '!src/components/**/index.ts',
+  ],
+  coverageThreshold: {
+    global: {
+      branches: 75,
+      functions: 80,
+      lines: 80,
+      statements: 80,
+    },
+  },
+};