--- a/jest.setup.js
+++ b/jest.setup.js
@@ -1,20 +1,23 @@
 // Jest setup file
 import '@testing-library/jest-dom';
 
-// Mock window.location using Object.defineProperty
-Object.defineProperty(window, 'location', {
-  writable: true,
-  value: {
-    href: '',
-    origin: 'http://localhost',
-    pathname: '/',
-    search: '',
-    hash: '',
-    assign: jest.fn(),
-    replace: jest.fn(),
-    reload: jest.fn(),
-  }
-});
+// Mock window.location - Kintsugi approach: Handle existing definition gracefully
+delete window.location;
+window.location = {
+  href: '',
+  origin: 'http://localhost',
+  pathname: '/',
+  search: '',
+  hash: '',
+  host: 'localhost',
+  hostname: 'localhost',
+  port: '',
+  protocol: 'http:',
+  assign: jest.fn(),
+  replace: jest.fn(),
+  reload: jest.fn(),
+  toString: () => 'http://localhost/'
+};
 
 // Mock Next.js router
 jest.mock('next/navigation', () => ({