… CICD_CONFIGURATION_UPDATES.md
… DEPENDENCY_UPGRADE_TEST_REPORT.md
… DEPLOYMENT_DASHBOARD.md
… E2E_TEST_SUMMARY.md
… PERFORMANCE_UPGRADE_REPORT.md
… README.md
… ROLLBACK_PROCEDURES.md
… SURGICAL_DEPLOYMENT_PLAN.md
… build-output.txt
… e2e/README.md
… e2e/financial-intelligence.test.ts
… e2e/generate-test-report.ts
… e2e/sales-intelligence.test.ts
… e2e/supply-chain-intelligence.test.ts
… package.json.backup
… performance-report.json
… performance-summary.txt
… scripts/performance-analysis.js
… scripts/staging-validation.sh
… src/__tests__/dependency-upgrade-test-suite.tsx
… src/__tests__/integration/dependency-upgrade.test.tsx
… src/__tests__/setup.ts
… tsconfig.json
… vercel.json
… vitest.config.ts